Added API calls for fetching and updating user roles.
